Removed Typescript
