+++ nodemon.json
@@ -1,6 +1,4 @@
 {
     "ignore": ["**/*.test.ts", "**/*.spec.ts", ".git", "node_modules"],
-    "ext": "ts",
     "exec": "npm run server"
-  }
-  
+}
