+++ package.json
@@ -1,17 +1,18 @@
 {
-    "name": "year-scale-static",
+    "name": "",
     "version": "0.0.0",
     "private": true,
     "scripts": {
         "start": "concurrently \"npm run server\" \"npm run client\"",
-        "server": "npx tsc && node ./bin/www",
-        "server:watch": "nodemon",
-        "dev": "concurrently \"npm run server:watch\" \"npm run client:watch\"",
-        "client:watch": "webpack-dev-server  --config webpack.dev.config.js --mode development --hot --open",
+        "server": "node ./bin/www",
+        "server:build": "babel src/backend/ --out-dir dist/",
+        "server:watch": "concurrently \"nodemon\" \"babel src/backend/ --out-dir dist/ -w\"",
         "client": "webpack-dev-server  --config webpack.dev.config.js --mode development --open",
-        "test": "echo \"Error: no test specified\" && exit 1",
-        "build": "npx tsc && webpack --config webpack.production.config.js --mode production",
-        "build:client": "webpack --config webpack.production.config.js --mode production"
+        "client:build": "webpack --config webpack.production.config.js --mode production",
+        "client:watch": "webpack-dev-server  --config webpack.dev.config.js --mode development --hot --open",
+        "build": "npm run client:build && npm run server:build",
+        "dev": "concurrently \"npm run server:watch\" \"npm run client:watch\"",
+        "test": "echo \"Error: no test specified\" && exit 1"
     },
     "dependencies": {
         "concurrently": "^5.3.0",
@@ -24,16 +25,12 @@
         "react-router-dom": "^5.1.2"
     },
     "devDependencies": {
-        "@babel/core": "^7.6.4",
+        "@babel/cli": "^7.13.14",
+        "@babel/core": "^7.13.15",
+        "@babel/node": "^7.13.13",
         "@babel/preset-env": "^7.6.3",
         "@babel/preset-react": "^7.6.3",
-        "@types/cookie-parser": "^1.4.2",
-        "@types/express": "^4.17.1",
-        "@types/morgan": "^1.7.37",
-        "@types/node": "^12.11.1",
-        "@types/react-dom": "^16.9.2",
-        "@types/react-router-dom": "^5.1.2",
-        "babel-loader": "^8.0.6",
+        "babel-loader": "^8.2.2",
         "css-loader": "^3.2.0",
         "debug": "~2.6.9",
         "html-webpack-plugin": "^3.2.0",
@@ -43,8 +40,6 @@
         "prettier": "^2.1.2",
         "source-map-loader": "^0.2.4",
         "style-loader": "^1.0.0",
-        "ts-loader": "^6.2.1",
-        "typescript": "^3.9.9",
         "webpack": "^4.46.0",
         "webpack-cli": "^3.3.9",
         "webpack-dev-server": "^3.9.0",
