--- src/backend/index.ts
@@ -1,20 +1,0 @@
-import express from 'express';
-import path from 'path';
-import cookieParser from 'cookie-parser';
-import logger from 'morgan';
-
-import indexRouter from './routes/index.js';
-import usersRouter from './routes/users.js';
-
-const index = express();
-
-index.use(logger('dev'));
-index.use(express.json());
-index.use(express.urlencoded({ extended: false }));
-index.use(cookieParser());
-index.use(express.static('public'));
-
-index.use('/', indexRouter);
-index.use('/api/users', usersRouter);
-
-module.exports = index;
