+++ src/client/components/App.jsx
@@ -1,9 +1,8 @@
-
 import React from "react";
 
 const App = () => (
     <div className="container">
-        <h1>Hello World, React!!!!!</h1>
+        <h1>Hello World, React!!!!</h1>
     </div>
 );
