+++ webpack.config.js
@@ -1,44 +1,39 @@
-const path = require('path');
+const path = require("path");
 const HtmlWebpackPlugin = require("html-webpack-plugin");
 const MiniCssExtractPlugin = require("mini-css-extract-plugin");
 
 module.exports = {
     entry: {
-        index: 'src/client/index.tsx'
+        index: "src/client/index.jsx",
     },
     output: {
         filename: "javascript/bundle.[hash].js",
-        path: path.resolve(__dirname, 'public')
+        path: path.resolve(__dirname, "public"),
     },
     resolve: {
-        modules: [__dirname, 'src', 'node_modules'],
-        extensions: ['*', '.js', '.jsx', '.tsx', '.ts'],
+        modules: [__dirname, "src", "node_modules"],
+        extensions: ["*", ".js", ".jsx", ".tsx", ".ts"],
     },
     module: {
         rules: [
             {
                 test: /\.jsx?$/,
                 exclude: /node_modules/,
-                loader: require.resolve('babel-loader')
+                loader: require.resolve("babel-loader"),
             },
             {
                 test: /\.css$/,
-                use: ['style-loader', 'css-loader']
+                use: ["style-loader", "css-loader"],
             },
-            {
-                test: /\.tsx?$/,
-                use: 'ts-loader',
-                exclude: /node_modules/
-            },
-        ]
+        ],
     },
-    plugins:  [
+    plugins: [
         new HtmlWebpackPlugin({
-            template: 'src/client/index.html',
-            filename: 'index.html'
+            template: "src/client/index.html",
+            filename: "index.html",
         }),
         new MiniCssExtractPlugin({
-            filename: 'styles/bundle.[hash].css'
+            filename: "styles/bundle.[hash].css",
         }),
-    ]
+    ],
 };
